# backend-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bal-Con Builders API",
    "version": "0.3.1",
    "description": "Construction management platform API (incremental exports, webhooks, analytics)."
  },
...
  "paths": {
...
    "/api/materials": {
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Create material",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterial"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "List materials",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Material"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders": {
      "post": {
        "summary": "Create work order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List work orders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List orders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analytics/distribution/materials": {
      "get": {
        "summary": "Get material distribution by field",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "category",
                "status"
              ],
              "default": "category"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        }
      },
...
      "Material": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "currentStock": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "reorderPoint": {
            "type": "number"
          },
          "unitCost": {
            "type": "number"
          },
          "markupPercentage": {
            "type": "number"
          },
          "sellingPrice": {
            "type": "number"
          },
          "supplierName": {
            "type": "string",
            "nullable": true
          },
          "supplierContact": {
            "type": "string",
            "nullable": true
          },
          "supplierEmail": {
            "type": "string",
            "nullable": true
          },
          "leadTimeDays": {
            "type": "integer"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "discontinued"
            ]
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "projectId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "quoteId": {
            "type": "integer",
            "nullable": true
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "in_production",
              "ready",
              "shipped",
              "delivered",
              "completed",
              "cancelled"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "taxAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "amountPaid": {
            "type": "number"
          },
          "items": {},
          "shippingAddress": {
            "nullable": true
          },
          "billingAddress": {
            "nullable": true
          },
          "paymentTerms": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "estimatedDelivery": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualDelivery": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shippedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateMaterial": {
        "type": "object",
        "required": [
          "name",
          "category",
          "unitOfMeasure",
          "unitCost",
          "markupPercentage",
          "sellingPrice",
          "leadTimeDays",
          "status",
          "currentStock",
          "minimumStock",
          "reorderPoint"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "currentStock": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "reorderPoint": {
            "type": "number"
          },
          "unitCost": {
            "type": "number"
          },
          "markupPercentage": {
            "type": "number"
          },
          "sellingPrice": {
            "type": "number"
          },
          "supplierName": {
            "type": "string"
          },
          "supplierContact": {
            "type": "string"
          },
          "supplierEmail": {
            "type": "string"
          },
          "leadTimeDays": {
            "type": "integer"
          },
          "location": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "discontinued"
            ]
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateOrder": {
        "type": "object",
        "required": [
          "projectId",
          "userId",
          "orderNumber",
          "subtotal",
          "taxAmount",
          "totalAmount",
          "amountPaid",
          "items"
        ],
        "properties": {
          "projectId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "quoteId": {
            "type": "integer"
          },
          "orderNumber": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "taxAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "amountPaid": {
            "type": "number"
          },
          "items": {},
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "in_production",
              "ready",
              "shipped",
              "delivered",
              "completed",
              "cancelled"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}